Change int/long

+++ src/main/java/ru/netology/stats/StatsService.java
@@ -4,21 +4,21 @@ package ru.netology.stats;
 public class StatsService {
 
     // сумма всех продаж
-    public int sumOfAllSales(int[] sales) {
-        int allSales = 0;
-        for (int sale : sales) {
+    public long sumOfAllSales(int[] sales) {
+        long allSales = 0;
+        for (long sale : sales) {
             allSales += sale;
         }
         return allSales;
     }
 
     // средняя сумма продаж в месяц
-    public int averageSalesAmount(int[] sales) {
-        int allSales = 0;
-        for (int sale : sales) {
+    public long averageSalesAmount(int[] sales) {
+        long allSales = 0;
+        for (long sale : sales) {
             allSales += sale;
         }
-        int averageSalesAmount = allSales / 12;
+        int averageSalesAmount = (int) (allSales / 12);
 
         return averageSalesAmount;
     }
@@ -52,7 +52,7 @@ public class StatsService {
 
     // количество месяцев, в которых продажи были ниже среднего
     public int salesAreBelowAverage(int[] sales) {
-        int averageSalesAmount = averageSalesAmount(sales);
+        int averageSalesAmount = (int) averageSalesAmount(sales);
         int belowAverage = 0;
 
         for (int sale : sales) {
@@ -65,7 +65,7 @@ public class StatsService {
 
     // количество месяцев, в которых продажи были выше среднего
     public int salesAreAboveAverage(int[] sales) {
-        int averageSalesAmount = averageSalesAmount(sales);
+        int averageSalesAmount = (int) averageSalesAmount(sales);
         int aboveAverage = 0;
 
         for (int sale : sales) {
